tested

+++ app/Http/Controllers/AlbumController.php
@@ -69,7 +69,10 @@ class AlbumController extends Controller
             $album->cover_image = pathinfo(
                 Storage::disk('images')->putFile('/', $request->file('file'))
             )['basename'];
-            Storage::disk('images')->delete($oldImg);
+            
+            if ($oldImg !== 'default_album.jpg') {
+                Storage::disk('images')->delete($oldImg);
+            }
         }
 
         if (!empty($request->name)) {
